+++ ckanext/notifications/logic/auth.py
@@ -1,10 +1,16 @@
-import ckan.plugins.toolkit as tk
 from ckan import types
 
 
 def notification_list_auth(context: types.Context, data_dict: types.DataDict) -> types.AuthResult:
     """Ensures users can only pull up their own personal notification rows."""
-    if tk.check_access("user_edit", context, data_dict):
+    requested_user = data_dict.get("user_id")
+    auth_user_obj = context.get("auth_user_obj")
+    authenticated_user = context.get("user")
+
+    if requested_user and auth_user_obj and requested_user in [auth_user_obj.id, auth_user_obj.name]:
+        return {"success": True}
+
+    if requested_user and authenticated_user == requested_user:
         return {"success": True}
 
     return {